backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=13,<15
pydantic==2.5.0
python-multipart==0.0.6
google-genai==1.40.0
protobuf>=4.25.3,<5
requests==2.32.3
beautifulsoup4==4.12.3
reportlab==4.2.2
python-dotenv==1.0.1
typing-extensions>=4.10,<5
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
litellm==1.52.11
